Fixed warning in directory iterator, on GCC 4.4.5 (Debian 4.4.4-8).
git-svn-id: file:///home/michael/svn/yuni/trunk@1193 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/core/io/directory/info/platform.cpp
@@ -241,7 +241,7 @@ namespace Directory
 
 		}
 
-		void push(const char* const str, unsigned int length)
+		void push(const char* const str, size_t length)
 		{
 			dirinfo.push_front();
 			dirinfo.front().parent.assign(str, length);
@@ -308,7 +308,7 @@ namespace Directory
 
 
 
-	IteratorData* IteratorDataCreate(const char* folder, unsigned int length, unsigned int flags)
+	IteratorData* IteratorDataCreate(const char* folder, size_t length, unsigned int flags)
 	{
 		if (length)
 		{
